test storage

+++ backend/models/EnergySource.cs
@@ -1,7 +1,9 @@
+using System;
+
 namespace Monocube.RoEnergy.Models;
 
 public sealed record EnergySource
 {
     public string Name { get; init; }
-    public float Capacity { get; init; }
+    public double Capacity { get; init; }
 }
